Convert to Python 3 syntax

+++ ensembl_prodinf/server_utils.py
@@ -3,7 +3,7 @@ import subprocess
 
 http_uri_regex = r"^(http){1}(s){0,1}(://){1}(.+){1}(:){1}(\d+){1}(/){1}(.+){0,1}$"
 uri_regex = r"^(mysql://){1}(.+){1}(:.+){0,1}(@){1}(.+){1}(:){1}(\d+){1}(/){1}$"
-db_uri_regex = r"^(mysql://){1}(.+){1}(:.+){0,1}(@){1}(.+){1}(:){1}(\d+){1}(/){1}(.+){1}$"    
+db_uri_regex = r"^(mysql://){1}(.+){1}(:.+){0,1}(@){1}(.+){1}(:){1}(\d+){1}(/){1}(.+){1}$"
 email_regex = r"^(.+){1}(@){1}(.+){1}$"
 
 """
@@ -23,26 +23,26 @@ def assert_mysql_uri(uri):
 def assert_mysql_db_uri(uri):
     """Check supplied URI matches MySQL database"""
     if not re.search(db_uri_regex, uri):
-        raise ValueError("MySQL database URL doesn't match pattern: mysql://user(:pass)@server:port/prod_db_name")   
+        raise ValueError("MySQL database URL doesn't match pattern: mysql://user(:pass)@server:port/prod_db_name")
 
 def assert_email(email):
     """Check supplied string is an email address"""
     if not re.search(email_regex, email):
         raise ValueError("Email doesn't match pattern: user@domain")
 
-def get_load(host=None): 
+def get_load(host=None):
     """Find load by on the supplied host by ssh (or on localhost if no host is supplied)
     """
     # load obtained from uptime
     status = run_process('uptime', process_uptime, host)
     return status
 
-def get_file_sizes(host=None, dir_name=None):     
+def get_file_sizes(host=None, dir_name=None):
     """Find file sizes in the supplied directory on the supplied host by ssh (or on localhost if no host is supplied)
     """
     # determine file sizes with du
     return run_process('"(cd ' + dir_name + ' && du -sm *)"', process_du, host)
- 
+
 def get_status(host=None, dir_name=None):
     """Base entry point for getting all status for a host, returned as a dict
     Arguments:
@@ -104,7 +104,7 @@ def process_ncores(status, line):
 def process_du(status, line):
     """Internal method to parse output of du and add to status hash"""
     elems = line.split()
-    status[elems[1]] = long(elems[0])
+    status[elems[1]] = int(elems[0])
 
 
 def run_process(command, function, host=None):
